first draft of registry with migrations
