+++ plugins/disk_sizes/01_db_init.rb
@@ -1,0 +1,14 @@
+# Disk_Size's migration file
+
+Sequel.migration do
+  transaction
+  change do
+    create_table(:measurementDiskSizes) do # Plugin's DB model
+      primary_key   :id
+      foreign_key   :node_resource
+      Time          :time, null: false
+      String        :node
+      Integer       :value
+    end
+  end
+end
